refactor(core): extract execute_post into handler/post.rs; handler.rs ≤500
Second of two post-PR-4c extractions. PR #111 split out execute_delete
+ blocking helpers into handler/delete.rs; this PR splits out
execute_post into handler/post.rs and brings handler.rs from 523 to
~395 production lines, finally below the 500-line ceiling.

After this PR, only `core/src/main.rs` (~847 production) remains
in violation of the 500-line bar. The grandfather clause AGENTS.md
section is rewritten to reflect that — handler.rs no longer
appears in the active reduction list, and the clause's exit
criterion drops to a single condition (main.rs ≤ 500).

## What moves

`core/src/handler/post.rs` (new, 165 lines):

- `pub(crate) async fn execute_post` — verbatim from handler.rs.
- Imports `super::etag_preview` since both PUT (still in handler.rs)
  and POST emit the same `sqlite_committed etag=...` aux trace line.

`core/src/handler.rs` (523 → 396 lines):

- `mod post;` + `pub(crate) use post::execute_post;` next to the
  delete re-export. No call-site changes in main.rs / pipeline.rs;
  the dispatcher and white-box tests keep their import path stable.
- Deleted execute_post body (134 lines).
- `etag_preview` lifted from `fn` to
  `pub(in crate::handler) fn` so post.rs can `use super::etag_preview;`.
  Visibility scoped to the handler module tree only — etag
  presentation is verb-handler-internal, not a crate-wide utility.

## Doc nits fixed (carry-over from PR #111 review)

- **AGENTS.md**: was "handler.rs (803 production)" — stale since
  PR #111 brought it to 523. Section rewritten to reflect post-this-PR
  state (handler.rs ~395, qualifies; main.rs is the sole remaining
  target).
- **handler/delete.rs module doc**: was "to keep handler.rs under
  the 500-line ceiling" — strictly false at the moment PR #111
  landed (handler.rs was 523, still 23 over). Reworded to
  "first of two post-PR-4c extractions that bring handler.rs back
  under the 500-line ceiling; the second (crate::handler::post)
  lands the same shape" — now accurate.

## Stats

- 3 files: handler.rs (-127 / +4), handler/delete.rs (+2 / -1
  doc), handler/post.rs (+165 new), AGENTS.md (-22 / +18).
- Net production: +43 lines (per-file doc + import set duplicated
  across the new module boundary; same shape as PR #111).
- `core/src/handler.rs` production: 523 → 396 (✓ ≤ 500).
- `core/src/handler/delete.rs`: 308 (✓).
- `core/src/handler/post.rs`: 165 (✓).
- `core/src/main.rs` production: unchanged at ~847 (PR 4d's job).

## Test plan

- [x] `cargo test` — 124 unit tests pass.
- [x] `cargo clippy --all-targets -- -D warnings` — clean.
- [x] `cargo fmt --check` — clean (after one auto-fmt pass).
- [x] `python sdk/tests/e2e_blackbox.py` — 233/233 checks pass.
- [ ] CI sdk-blackbox on Windows — same flake class as before;
      PR #109's CoAP timeout fix is in master.

## What's next

PR 4d: split main.rs's route table + axum middleware + env helpers
+ `Core::new`/`main()` startup into `route.rs` / `middleware.rs` /
`config.rs`. Once main.rs ≤ 500 production lines, the grandfather
clause retires entirely.
