+++ core/src/handler.rs
@@ -35,7 +35,9 @@
 //! can be removed at that point.
 
 mod delete;
+mod post;
 pub(crate) use delete::execute_delete;
+pub(crate) use post::execute_post;
 
 use std::sync::atomic::Ordering;
 
@@ -380,144 +382,15 @@ pub(crate) async fn execute_put(
     Phase::CommittedWrite((status, to_header_map(resp_headers), "").into_response())
 }
 
-pub(crate) async fn execute_post(
-    headers: HeaderMap,
-    body: Bytes,
-    tier: auth::Tier,
-    world: String,
-    core: &Core,
-    trace: &TraceCtx,
-) -> Phase {
-    // POST appends to an existing world. PUT is the create/replace
-    // path — POST returns 404 if the world is absent and never
-    // updates X-Meta-* headers (which are PUT-only).
-    if !can_write(&world, tier) {
-        let gate = if needs_write_approve(&world) {
-            AuthGate::WriteApprove
-        } else {
-            AuthGate::Write
-        };
-        return Phase::Error {
-            resp: unauthorized("write requires token; system worlds need approve token"),
-            reason: ErrorReason::Auth(gate),
-        };
-    }
-    let _write_guard = core.acquire_world_lock(&world).await;
-    trace.emit_aux("lock_acquired");
-    if let Err(resp) = hs::check_write_preconditions(core, &world, &headers) {
-        let reason = if resp.status() == StatusCode::PRECONDITION_FAILED {
-            ErrorReason::PreconditionFailed
-        } else {
-            ErrorReason::StorageRead
-        };
-        return Phase::Error { resp, reason };
-    }
-    let Some((body_len, content_type, stored_headers)) = (match core.world_metadata(&world) {
-        Ok(meta) => meta,
-        Err(e) => {
-            return Phase::Error {
-                resp: storage_error("storage metadata", e),
-                reason: ErrorReason::StorageRead,
-            };
-        }
-    }) else {
-        return Phase::Error {
-            resp: not_found(),
-            reason: ErrorReason::NotFound,
-        };
-    };
-    let Some(projected_len) = body_len.checked_add(body.len()) else {
-        return Phase::Error {
-            resp: payload_too_large(core.max_world_bytes),
-            reason: ErrorReason::PayloadTooLarge,
-        };
-    };
-    if projected_len > core.max_world_bytes {
-        return Phase::Error {
-            resp: payload_too_large(core.max_world_bytes),
-            reason: ErrorReason::PayloadTooLarge,
-        };
-    }
-    let new_etag = if store::is_persistent(&world) {
-        // POST adds bytes on top; existing bytes already counted.
-        // prev_len = 0 so the reservation only sizes the delta.
-        if let Some(quota) = core.max_storage_bytes {
-            let used = core.storage_body_bytes.load(Ordering::Relaxed);
-            trace.emit_aux_kv("quota_check", &format!("used={used} quota={quota}"));
-        }
-        if let Err(boxed) = core.reserve_storage(0, body.len()) {
-            return Phase::Error {
-                resp: *boxed,
-                reason: ErrorReason::QuotaExceeded,
-            };
-        }
-        match world::append_with_audit(
-            &core.data,
-            &world,
-            &body,
-            &content_type,
-            &stored_headers,
-            &core.hmac_key,
-        ) {
-            Ok(Some((_result, h))) => {
-                let etag = hs::hmac_etag(&h);
-                trace.emit_aux_kv("sqlite_committed", &format!("etag={}", etag_preview(&etag)));
-                etag
-            }
-            Ok(None) => {
-                // World disappeared between metadata read and append.
-                core.rollback_storage_reservation(0, body.len());
-                return Phase::Error {
-                    resp: not_found(),
-                    reason: ErrorReason::NotFound,
-                };
-            }
-            Err(e) => {
-                core.rollback_storage_reservation(0, body.len());
-                let reason = if is_insufficient_storage_error(&e) {
-                    ErrorReason::InsufficientStorage
-                } else {
-                    ErrorReason::StorageWriteAudit
-                };
-                return Phase::Error {
-                    resp: storage_error("storage/audit", e),
-                    reason,
-                };
-            }
-        }
-    } else {
-        match core
-            .mem
-            .append_with_quota(&world, &body, core.max_memory_bytes)
-        {
-            Ok(Some(result)) => {
-                let etag = format!("sha256-{}", result.body_sha256_after);
-                trace.emit_aux_kv("sqlite_committed", &format!("etag={}", etag_preview(&etag)));
-                etag
-            }
-            Ok(None) => {
-                return Phase::Error {
-                    resp: not_found(),
-                    reason: ErrorReason::NotFound,
-                };
-            }
-            Err(store::MemoryQuotaError { quota, .. }) => {
-                return Phase::Error {
-                    resp: payload_too_large(quota),
-                    reason: ErrorReason::PayloadTooLarge,
-                };
-            }
-        }
-    };
-    core.notify("POST", &world, &new_etag);
-    trace.emit_aux("notify_sent");
-    let resp_headers = [(header::ETAG, hs::etag_header(&new_etag))];
-    Phase::CommittedWrite((StatusCode::OK, resp_headers, "").into_response())
-}
-
 /// First 16 chars of an etag string for compact aux trace lines.
 /// Etags are HMAC-SHA256 hex (64 chars) or `sha256-<64 chars>`; a
 /// 16-char prefix is enough to disambiguate while staying readable.
-fn etag_preview(etag: &str) -> String {
+///
+/// Visibility is `pub(in crate::handler)` so the sibling `post.rs`
+/// module can reuse it (PUT and POST both emit the same
+/// `sqlite_committed etag=...` aux line). Not visible outside
+/// `crate::handler` — etag presentation is a verb-handler concern,
+/// not a crate-wide utility.
+pub(in crate::handler) fn etag_preview(etag: &str) -> String {
     etag.chars().take(16).collect()
 }
